+++ .gitignore
@@ -2,3 +2,4 @@
 .doc
 node_modules
 coverage/
+~*
